Replace STUB functionality
Replace FORCE_UNINIMPLEMENTED mechanism with STUB_API, which rather than
returning unimplemented and falling back on --stub-unimplemented to
replace them, replaces the functionality. Remove --stub-unimplemented
entirely and revert the zeCall() arguments to the output of a function
call rather than the function and its arguments (which for stubbed calls
becomes a dictionary describing the call).