Fix bug in sysman power CTS
Bug is in one of power cts, where when pBurstInitial is true
we are setting pBurstSet.enabled to false.
Problem occurs when while setting new burst power, we are
expecting new burst power values to take effect.
But those values wont take effect as we had disabled pBurstSet.
This change fixes that.

Signed-off-by: Jitendra Sharma <jitendra.sharma@intel.com>