fix: context creation in zeIPCEventTests group
The tests use two processes to exercise the IPC features: the parent
process creates an event-pool and sends all needed handlers to the child
process via socket. The child, in turn, uses the received data to open
the same event-pool and then use it.

When creating the event-pool, the parent can pass the list of associated
devices as an argument to `zeEventPoolCreate()`. On the other hand, when
the child opens the same event-pool, the `zeEventPoolOpenIpcHandle()`
does not accept any list of devices to specify. Instead, under the hood,
the list of devices is derived from the `context` instance.

In addition to that, so far the context is created using
`zeContextCreate()` and then the `context` instance in the child process
was associated with the root-device (instead of its subdevice[0]!).

Then, in NEO the method `DeviceImp::isImplicitScalingCapable()` returns
`true` for root-device and `false` for a subdevice. The method is used
by both the parent process (`EventPool::initialize()`) and the child
process (`EventPool::openEventPoolIpcHandle()`) and the different values
returned cause different results computed by
`DeviceImp::getEventMaxPacketCount()`. This, in turn, leads to an error
when opening the event-pool because a respective check fails.

Fix the problem in child process by using `zeContextCreateEx()`, which
accepts additional parameters - the list of devices (here,
a subdevice[0] only) a context should be associated with.

Related-To: NEO-9837
Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>