fix: avoid out-of-boundary access by `profile_add` kernel
The kernel uses `get_global_id()` to index the `values_out[]` array.
The index is therefore related to "global work-group size" (GWS) and
belongs to `[0, GWS)`.

Adjust the GWS so that the kernel does not reach out-of-boundary while
accessing each element of the array.
- set the "local work-group size" (LWS) to 1 with `set_group_size()`
- adjust the GWS by setting `group_count = elems_nb` when
  (`GWS = LWS * group_count`)`

Fix other tests in the same manner.

Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>