fix: do not access UMD-managed USM-allocation while in-use
With UMD-managed migrations, it is illegal to access an USM-allocation
from CPU side while it is still in use on GPU side. In such scenario
a CPU-side access triggers data fetch from GPU-side and then any
subsequent GPU-side modification will not be observable on the CPU-side.
With UMD-managed migrations there is no way to tell the GPU that it
should fetch the allocation again from CPU-side before modifying it.

Therefore the `EXPECT_NE` statements can only be called when concurrent
access to USM-allocations is supported.

Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>